requirements.txt
python-telegram-bot>=21.0
gspread>=6.0
google-auth>=2.0
httpx>=0.27
python-dotenv>=1.0
Pillow>=10.0
pytesseract>=0.3
